Solving Intersect Ray Sphere range issue

+++ sphere.cpp
@@ -1,6 +1,6 @@
 #include "sphere.h"
 
-float Sphere::getRadius()
+double Sphere::getRadius()
 {
     return radius;
 }
